app/dcl_engine/requirements.txt
duckdb>=1.0.0
fastapi==0.119.0
google-generativeai==0.8.5
pinecone==3.1.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
uvicorn==0.37.0
pandas
pyyaml
flask
jinja2
requests
